+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 0.7.0 - unreleased
 * Add feature/test for trait constant and associated type as trait field
 * Add feature/test for trait generics, including both trait_bound and where_clause styles
+* Add feature/test for original trait parents(bound) of the target trait
 
 ## 0.6.0- March 9, 2024
 * Refine tests